refactor: centralize common value objects
