+++ src/backend/src/routes/application/use-cases/get-route-details.ts
@@ -1,6 +1,6 @@
 import { RouteRepository } from '../../domain/repositories/route-repository';
 import { Route } from '../../domain/entities/route-entity';
-import { UUID } from '../../domain/value-objects/uuid-value-object';
+import { UUID } from '../../../shared/domain/value-objects/uuid-value-object';
 
 export class GetRouteDetailsUseCase {
   constructor(private repository: RouteRepository) {}
