+++ src/backend/src/routes/domain/events/route-generated.test.ts
@@ -1,6 +1,6 @@
 import { RouteGeneratedEvent } from './route-generated';
 import { Route } from '../entities/route-entity';
-import { UUID } from '../value-objects/uuid-value-object';
+import { UUID } from '../../../shared/domain/value-objects/uuid-value-object';
 import { DistanceKm } from '../value-objects/distance-value-object';
 import { Duration } from '../value-objects/duration-value-object';
 import { Path } from '../value-objects/path-value-object';
