+++ src/backend/src/routes/infrastructure/dynamodb/dynamo-route-repository.test.ts
@@ -1,5 +1,5 @@
 import { Route } from "../../domain/entities/route-entity";
-import { UUID } from "../../domain/value-objects/uuid-value-object";
+import { UUID } from "../../../shared/domain/value-objects/uuid-value-object";
 import { DistanceKm } from "../../domain/value-objects/distance-value-object";
 import { Duration } from "../../domain/value-objects/duration-value-object";
 import { Path } from "../../domain/value-objects/path-value-object";
