+++ src/backend/src/routes/infrastructure/in-memory/in-memory-route-repository.test.ts
@@ -1,6 +1,6 @@
 import { InMemoryRouteRepository } from "./in-memory-route-repository";
 import { Route } from "../../domain/entities/route-entity";
-import { UUID } from "../../domain/value-objects/uuid-value-object";
+import { UUID } from "../../../shared/domain/value-objects/uuid-value-object";
 import { DistanceKm } from "../../domain/value-objects/distance-value-object";
 import { Duration } from "../../domain/value-objects/duration-value-object";
 import { Path } from "../../domain/value-objects/path-value-object";
