+++ src/backend/src/routes/infrastructure/in-memory/in-memory-route-repository.ts
@@ -1,6 +1,6 @@
 import { Route } from "../../domain/entities/route-entity";
 import { RouteRepository } from "../../domain/repositories/route-repository";
-import { UUID } from "../../domain/value-objects/uuid-value-object";
+import { UUID } from "../../../shared/domain/value-objects/uuid-value-object";
 
 export class InMemoryRouteRepository implements RouteRepository {
   private routes = new Map<string, Route>();
