+++ src/backend/src/routes/interfaces/http/page-router.test.ts
@@ -45,7 +45,7 @@ jest.mock("../appsync-client", () => ({
 
 import { handler } from "./page-router";
 import { Route } from "../../domain/entities/route-entity";
-import { UUID } from "../../domain/value-objects/uuid-value-object";
+import { UUID } from "../../../shared/domain/value-objects/uuid-value-object";
 import { DistanceKm } from "../../domain/value-objects/distance-value-object";
 import { Duration } from "../../domain/value-objects/duration-value-object";
 import { Path } from "../../domain/value-objects/path-value-object";
