+++ src/backend/src/shared/domain/value-objects/email-value-object.ts
