+++ src/backend/src/shared/domain/value-objects/uuid-value-object.ts
