+++ src/backend/src/users/application/use-cases/get-user-profile.ts
@@ -1,4 +1,4 @@
-import { Email } from "../../../routes/domain/value-objects/email-value-object";
+import { Email } from "../../../shared/domain/value-objects/email-value-object";
 import { UserProfile } from "../../domain/entities/user-profile";
 import { UserStateRepository } from "../../domain/repositories/user-state-repository";
