+++ src/backend/src/users/application/use-cases/get-user-profile.usecase.test.ts
@@ -1,6 +1,6 @@
 import { GetUserProfileUseCase } from './get-user-profile';
 import { UserStateRepository } from '../../domain/repositories/user-state-repository';
-import { Email } from '../../../routes/domain/value-objects/email-value-object';
+import { Email } from '../../../shared/domain/value-objects/email-value-object';
 
 describe('GetUserProfileUseCase', () => {
   it('creates profile when missing', async () => {
