+++ src/backend/src/users/domain/entities/user-profile.ts
@@ -1,5 +1,5 @@
 
-import { Email } from '../../../routes/domain/value-objects/email-value-object';
+import { Email } from '../../../shared/domain/value-objects/email-value-object';
 import { Age } from '../value-objects/age-value-object';
 import { DistanceUnit } from '../value-objects/distance-unit-value-object';
